requirements.txt
pytest==9.0.2
selenium==4.39.0
webdriver-manager==4.0.2
pytest-html==4.1.1
allure-pytest==2.15.3
pytest-xdist==3.8.0